treewide: Update rust-toolchain to 2023-03-07
Bump the rust-toolchain to nightly-2023-03-07.

Basically all the changes here come from automatically fixing clippy
lints with `cargo clippy --fix`, with a couple of exceptions:

* A few places were wrapping `async fn`s with `#[track_caller]`, which
  has apparently always been a no-op - these have been removed
* A few places were doing `let _ = ...` on a future without `await`ing
  that future, which unintentionally would have never actually run that
  future. This has been fixed to actually `await` the future.

Reviewed-on: https://gerrit.readyset.name/c/readyset/+/4661
Tested-by: Buildkite CI

+++ dataflow-expression/src/eval/builtins.rs
@@ -150,7 +150,7 @@ where
             .unwrap()
             .with_month(1)
             .unwrap()
-            .num_days_from_ce() as i32;
+            .num_days_from_ce();
         let mut weekday = if monday_first {
             time.weekday().num_days_from_monday()
         } else {
@@ -164,8 +164,8 @@ where
             }
             week_year = true;
             year -= 1;
-            days = days_in_year(year) as i32;
-            first_daynr -= days as i32;
+            days = days_in_year(year);
+            first_daynr -= days;
             weekday = (weekday + 53 * 7 - days) % 7;
         }
